Cherry pick FTR test changes from #236326
