+++ x-pack/platform/test/fleet_api_integration/apis/agents/migrate.ts
@@ -8,6 +8,7 @@ import expect from '@kbn/expect';
 import { AGENTS_INDEX, AGENT_ACTIONS_INDEX } from '@kbn/fleet-plugin/common';
 import type { FtrProviderContext } from '../../../api_integration/ftr_provider_context';
 import { enableActionSecrets } from '../../helpers';
+import { checkBulkAgentAction } from './helpers';
 
 export default function (providerContext: FtrProviderContext) {
   const { getService } = providerContext;
@@ -473,27 +474,7 @@ export default function (providerContext: FtrProviderContext) {
           .expect(200);
 
         const actionId = body.actionId;
-
-        await new Promise((resolve, reject) => {
-          let attempts = 0;
-          const intervalId = setInterval(async () => {
-            if (attempts > 5) {
-              clearInterval(intervalId);
-              reject(new Error('action timed out'));
-            }
-            ++attempts;
-            const {
-              body: { items: actionStatuses },
-            } = await supertest.get(`/api/fleet/agents/action_status`).set('kbn-xsrf', 'xxx');
-            const action = actionStatuses?.find((a: any) => a.actionId === actionId);
-            if (action && action.nbAgentsActioned === action.nbAgentsActionCreated) {
-              clearInterval(intervalId);
-              resolve({});
-            }
-          }, 3000);
-        }).catch((e) => {
-          throw e;
-        });
+        await checkBulkAgentAction(supertest, actionId);
       });
     });
   });
